added dtos, modified dao and entities, updated gitignore

+++ Tikito/src/main/java/com/sunbeam/tikito/daos/BookingDao.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.sunbeam.tikito.entity.BookingEntity;
 
-public interface BookingDao extends JpaRepository<BookingEntity, Integer>
+public interface BookingDao extends JpaRepository<BookingEntity, Long>
 {
 	List<BookingEntity> findByUserUserId(int userId);
 	List<BookingEntity> findByBookingIdAndUserUserId(int bookingId, int userId);
